Run codec diagnostics at startup and show GPU/CPU decode badges
